# Conflicts:
#	wifi4eu-common/src/main/java/wifi4eu/wifi4eu/common/dto/model/RegistrationDTO.java
#	wifi4eu-portal/wifi4eu-portal-service/src/main/java/wifi4eu/wifi4eu/service/registration/RegistrationService.java
#	wifi4eu-portal/wifi4eu-portal-web/pom.xml

+++ wifi4eu-portal/wifi4eu-portal-service/src/main/java/wifi4eu/wifi4eu/service/municipality/MunicipalityService.java
@@ -57,7 +57,8 @@ public class MunicipalityService {
             if (mayor != null) {
                 mayorService.deleteMayor(mayor.getId());
             }
-            for (RegistrationDTO registration : registrationService.getRegistrationsByMunicipalityId(municipalityDTO.getId())) {
+            RegistrationDTO registration = registrationService.getRegistrationByMunicipalityId(municipalityDTO.getId());
+            if (registration != null) {
                 for (ApplicationDTO application : applicationService.getApplicationsByRegistrationId(registration.getId())) {
                     applicationService.deleteApplication(application.getId());
                 }
